Re-Arranging GitHub Directory
pls no break everything
