--- project/proposal/Scrap2
@@ -1,1 +1,0 @@
-Scrap2
