--- project/scrap01.py
@@ -1,1 +1,0 @@
-# This is the initial scrap file for spare code and testing!
