--- test_data/sample_output.py
@@ -1,15 +1,0 @@
-#format: return (["City", "Cuisine", "Price", "Rating", "# Restaurants"], result_table)
-
-sample_output = (['City', 'Cuisine', 'Price', 'Rating', '# Restaurants'],
- [('Chicago', 'Seafood', 1.36, 3.83, 1),
-  ('Chicago', 'BBQ', 1.19, 3.69, 1),
-  ('Chicago', 'Cafe', 1.19, 3.69, 1),
-  ('Chicago', 'Sushi', 1.36, 3.515, 2),
-  ('Chicago', 'Thai', 1.12, 3.22, 1),
-  ('Chicago', 'Steak', 1.84, 3.2, 1),
-  ('Chicago', 'Mediterranean', 4.12, 3.17, 1),
-  ('Chicago', 'Mexican', 2.17, 3.095, 2),
-  ('Chicago', 'Indian', 2.17, 2.97, 1),
-  ('Chicago', 'Sandwich', 2.22, 2.26, 1)])
-
-
