--- yelp_data/city_ratings.csv
@@ -1,56 +1,0 @@
-City,Average Rating
-albuquerque,3.514517506404782
-arlington,3.4711309523809524
-atlanta,3.418818592640413
-austin,3.505167958656331
-baltimore,3.465162311955661
-boston,3.497507585609016
-buffalo,3.5662952646239554
-charlotte,3.4238767174155216
-chicago,3.554373246024322
-cleveland,3.5059769167353667
-columbus,3.435959595959596
-dallas,3.500550834120239
-dayton,3.505111524163569
-detroit,3.462028247279463
-fort wayne,3.6545012165450124
-fort worth,3.3598165592464055
-fresno,3.3985865724381625
-honolulu,3.5810008481764206
-houston,3.408230152949745
-indianapolis,3.472513638271087
-jacksonville,3.541801510248112
-kansas city,3.412615255658005
-las vegas,3.426852513700262
-long beach,3.541673085811123
-los angeles,3.7068091899770943
-louisville,3.5179224579370887
-memphis,3.47249434815373
-mesa,3.4693215339233037
-miami,3.511379800853485
-milwaukee,3.438082191780822
-minneapolis,3.4348471356485604
-nashville,3.4686418685121105
-new orleans,3.5850886917960088
-new york,3.642176159429204
-oakland,3.6098823529411765
-oklahoma city,3.5301992528019923
-omaha,3.481874447391689
-philadelphia,3.550259965337955
-phoenix,3.461869618696187
-pittsburgh,3.5434782608695654
-portland,3.611607142857143
-raleigh,3.4768062397372743
-sacramento,3.440344062153163
-salt lake city,3.4122340425531914
-san antonio,3.3857581967213113
-san diego,3.629283489096573
-san francisco,3.6254528272168844
-san jose,3.4604110813226097
-st louis,3.4744809688581313
-st paul,3.50291686922703
-tampa,3.5746617466174664
-tucson,3.445800430725054
-tulsa,3.586713995943205
-virginia beach,3.434843531060252
-washington dc,3.3978142076502733
