--- yelp_data/city_ratings_adjusted.csv
@@ -1,56 +1,0 @@
-City,Average Rating
-albuquerque,3.5001368129964154
-arlington,3.500136812996566
-atlanta,3.500136812996522
-austin,3.5001368129963946
-baltimore,3.500136812996494
-boston,3.5001368129961823
-buffalo,3.500136812996391
-charlotte,3.5001368129964643
-chicago,3.500136812996404
-cleveland,3.500136812996448
-columbus,3.5001368129964225
-dallas,3.5001368129962853
-dayton,3.5001368129963786
-detroit,3.500136812996241
-fort wayne,3.500136812996418
-fort worth,3.500136812996507
-fresno,3.5001368129963817
-honolulu,3.5001368129964336
-houston,3.5001368129964794
-indianapolis,3.500136812996348
-jacksonville,3.500136812996437
-kansas city,3.500136812996429
-las vegas,3.5001368129963923
-long beach,3.5001368129964865
-los angeles,3.5001368129964137
-louisville,3.5001368129965043
-memphis,3.500136812996423
-mesa,3.500136812996445
-miami,3.500136812996415
-milwaukee,3.5001368129964026
-minneapolis,3.5001368129964137
-nashville,3.5001368129963804
-new orleans,3.5001368129963812
-new york,3.5001368129965074
-oakland,3.5001368129963217
-oklahoma city,3.5001368129964168
-omaha,3.500136812996421
-philadelphia,3.5001368129964376
-phoenix,3.5001368129963533
-pittsburgh,3.5001368129963932
-portland,3.5001368129966104
-raleigh,3.5001368129964074
-sacramento,3.5001368129964323
-salt lake city,3.5001368129964003
-san antonio,3.5001368129964927
-san diego,3.5001368129964567
-san francisco,3.500136812996371
-san jose,3.500136812996632
-st louis,3.500136812996403
-st paul,3.5001368129965056
-tampa,3.500136812996407
-tucson,3.50013681299643
-tulsa,3.5001368129964274
-virginia beach,3.5001368129964376
-washington dc,3.500136812996508
