--- yelp_data/cuisines_tags.csv
@@ -1,149 +1,0 @@
-Afghan (afghani)
-African (african)
-Senegalese (senegalese)
-South African (southafrican)
-American (New) (newamerican)
-American (Traditional) (tradamerican)
-Arabian (arabian)
-Argentine (argentine)
-Armenian (armenian)
-Asian Fusion (asianfusion)
-Australian (australian)
-Austrian (austrian)
-Bangladeshi (bangladeshi)
-Barbeque (bbq)
-Basque (basque)
-Belgian (belgian)
-Brasseries (brasseries)
-Brazilian (brazilian)
-Breakfast & Brunch (breakfast_brunch)
-British (british)
-Buffets (buffets)
-Burgers (burgers)
-Burmese (burmese)
-Cafes (cafes)
-Themed Cafes (themedcafes)
-Cafeteria (cafeteria)
-Cajun/Creole (cajun)
-Cambodian (cambodian)
-Caribbean (caribbean)
-Dominican (dominican)
-Haitian (haitian)
-Puerto Rican (puertorican)
-Trinidadian (trinidadian)
-Catalan (catalan)
-Cheesesteaks (cheesesteaks)
-Chicken Shop (chickenshop)
-Chicken Wings (chicken_wings)
-Chinese (chinese)
-Cantonese (cantonese)
-Dim Sum (dimsum)
-Hainan (hainan)
-Shanghainese (shanghainese)
-Szechuan (szechuan)
-Comfort Food (comfortfood)
-Creperies (creperies)
-Cuban (cuban)
-Czech (czech)
-Delis (delis)
-Diners (diners)
-Dinner Theater (dinnertheater)
-Ethiopian (ethiopian)
-Fast Food (hotdogs)
-Filipino (filipino)
-Fish & Chips (fishnchips)
-Fondue (fondue)
-Food Court (food_court)
-Food Stands (foodstands)
-French (french)
-Mauritius (mauritius)
-Reunion (reunion)
-Gastropubs (gastropubs)
-German (german)
-Gluten-Free (gluten_free)
-Greek (greek)
-Guamanian (guamanian)
-Halal (halal)
-Hawaiian (hawaiian)
-Himalayan/Nepalese (himalayan)
-Honduran (honduran)
-Hong Kong Style Cafe (hkcafe)
-Hot Dogs (hotdog)
-Hot Pot (hotpot)
-Hungarian (hungarian)
-Iberian (iberian)
-Indian (indpak)
-Indonesian (indonesian)
-Irish (irish)
-Italian (italian)
-Calabrian (calabrian)
-Sardinian (sardinian)
-Tuscan (tuscan)
-Japanese (japanese)
-Conveyor Belt Sushi (conveyorsushi)
-Izakaya (izakaya)
-Japanese Curry (japacurry)
-Ramen (ramen)
-Teppanyaki (teppanyaki)
-Kebab (kebab)
-Korean (korean)
-Kosher (kosher)
-Laotian (laotian)
-Latin American (latin)
-Colombian (colombian)
-Salvadoran (salvadoran)
-Venezuelan (venezuelan)
-Live/Raw Food (raw_food)
-Malaysian (malaysian)
-Mediterranean (mediterranean)
-Falafel (falafel)
-Mexican (mexican)
-Tacos (tacos)
-Middle Eastern (mideastern)
-Egyptian (egyptian)
-Lebanese (lebanese)
-Modern European (modern_european)
-Mongolian (mongolian)
-Moroccan (moroccan)
-New Mexican Cuisine (newmexican)
-Nicaraguan (nicaraguan)
-Noodles (noodles)
-Pakistani (pakistani)
-Pan Asian (panasian)
-Persian/Iranian (persian)
-Peruvian (peruvian)
-Pizza (pizza)
-Polish (polish)
-Pop-Up Restaurants (popuprestaurants)
-Portuguese (portuguese)
-Poutineries (poutineries)
-Russian (russian)
-Salad (salad)
-Sandwiches (sandwiches)
-Scandinavian (scandinavian)
-Scottish (scottish)
-Seafood (seafood)
-Singaporean (singaporean)
-Slovakian (slovakian)
-Soul Food (soulfood)
-Soup (soup)
-Southern (southern)
-Spanish (spanish)
-Sri Lankan (srilankan)
-Steakhouses (steak)
-Supper Clubs (supperclubs)
-Sushi Bars (sushi)
-Syrian (syrian)
-Taiwanese (taiwanese)
-Tapas Bars (tapas)
-Tapas/Small Plates (tapasmallplates)
-Tex-Mex (tex-mex)
-Thai (thai)
-Turkish (turkish)
-Ukrainian (ukrainian)
-Uzbek (uzbek)
-Vegan (vegan)
-Vegetarian (vegetarian)
-Vietnamese (vietnamese)
-Waffles (waffles)
-Wraps (wraps)
