+++ yelp_data_assembly/starbucks_index.csv
@@ -1,0 +1,57 @@
+,City,Median Distance
+0,honolulu,0.03339150244309699
+1,miami,0.14242500981480086
+2,tampa,0.32465843310643305
+3,san antonio,0.36324401337682494
+4,new orleans,0.10371506449101939
+5,houston,0.3933192280098475
+6,jacksonville,0.1847491977107278
+7,austin,0.12341303367458928
+8,el paso,0.2417703554174012
+9,tucson,0.1351886573689471
+10,fort worth,0.23942829592613926
+11,san diego,0.110482301697517
+12,dallas,0.5176910349489224
+13,mesa,0.4110718600310712
+14,phoenix,0.5311653077191913
+15,atlanta,0.35259087674469547
+16,long beach,0.11243594712447123
+17,los angeles,0.18745364568783837
+18,albuquerque,0.15415496093571657
+19,memphis,0.2192677568261533
+20,charlotte,0.25099583378380447
+21,oklahoma city,0.21386409169299264
+22,raleigh,0.5645758483919439
+23,nashville,0.1734737119591012
+24,tulsa,0.32775416893749054
+25,las vegas,0.3954325306065184
+26,fresno,0.09681028138202319
+27,virginia beach,0.14052217801437886
+28,san jose,0.08993615678543937
+29,san francisco,0.06707456393838006
+30,oakland,0.09786786161055204
+31,sacramento,0.08443608702738065
+32,louisville,0.2134618709493967
+33,st louis,0.6356392999412528
+34,colorado springs,0.22151876805627174
+35,washington dc,0.10663066459045437
+36,arlington,0.08332804902360601
+37,baltimore,0.08640811005053274
+38,kansas city,0.9784717215630337
+39,indianapolis,0.40686133604370284
+40,denver,0.2824603384522651
+41,philadelphia,0.10638161691205382
+42,columbus,0.38515719786013786
+43,pittsburgh,0.27221693952704784
+44,new york,0.10539471592263164
+45,cleveland,0.14990331297494983
+46,omaha,0.2280611153139453
+47,chicago,0.13872551361195518
+48,milwaukee,0.14695002866407458
+49,detroit,0.48672985457481077
+50,boston,0.09831718190214267
+51,buffalo,0.1629276839837416
+52,minneapolis,0.3178584482253649
+53,st paul,0.12564368306157112
+54,portland,0.19391541829014255
+55,seattle,0.13421542012779097
